refactor:디렉터리 및 주석변경
