+++ src/main/java/com/plog/plogbackend/domain/Member/enums/Role.java
@@ -1,4 +1,4 @@
-package com.plog.plogbackend.domain.Member;
+package com.plog.plogbackend.domain.Member.enums;
 
 public enum Role {
   ROLE_USER, // 정상 회원가입 완료 회원
