+++ src/main/java/com/plog/plogbackend/domain/Member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.plog.plogbackend.domain.Member;
+package com.plog.plogbackend.domain.Member.repository;
 
+import com.plog.plogbackend.domain.Member.Member;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
