+++ src/main/resources/application-local.yml
@@ -7,7 +7,7 @@ spring:
   security:
     front:
       cors:
-        allowed-origins: "http://localhost:3000" # 프론트 CORS 허용 주소
+        allowed-origins: "http://localhost:8080" # 프론트 CORS 허용 주소
       frontend-url: http://localhost:8080
 
   servlet:
